refactor code to reduce widget nest in weather widget
