--- lib/src/widgets/line_chart.dart
@@ -1,21 +1,0 @@
-import 'package:charts_flutter/flutter.dart' as charts;
-import 'package:flutter/material.dart';
-
-class PointsLineChart extends StatelessWidget {
-  final List<charts.Series> seriesList;
-  final bool animate;
-
-  PointsLineChart(this.seriesList, {this.animate});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-        padding: const EdgeInsets.all(20.0),
-        child: charts.TimeSeriesChart(seriesList,
-            animate: animate,
-            animationDuration: Duration(milliseconds: 500),
-            primaryMeasureAxis: new charts.NumericAxisSpec(
-                tickProviderSpec: new charts.BasicNumericTickProviderSpec(
-                    zeroBound: false))));
-  }
-}
