+++ lib/src/widgets/weather_widget.dart
@@ -1,30 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_weather/main.dart';
 import 'package:flutter_weather/src/model/weather.dart';
-import 'package:flutter_weather/src/widgets/current_conditions.dart';
-import 'package:flutter_weather/src/widgets/empty_widget.dart';
 import 'package:flutter_weather/src/widgets/forecast_horizontal_widget.dart';
-import 'package:flutter_weather/src/widgets/line_chart.dart';
 import 'package:flutter_weather/src/widgets/value_tile.dart';
+import 'package:flutter_weather/src/widgets/weather_swipe_pager.dart';
 import 'package:intl/intl.dart';
-import 'package:flutter_swiper/flutter_swiper.dart';
-import 'package:charts_flutter/flutter.dart' as charts;
 
 class WeatherWidget extends StatelessWidget {
   final Weather weather;
-  final temperatureData;
 
-  WeatherWidget({this.weather})
-      : temperatureData = [
-          new charts.Series<Weather, DateTime>(
-            id: 'Temperature',
-            colorFn: (_, __) => charts.MaterialPalette.blue.shadeDefault,
-            domainFn: (Weather weather, _) =>
-                DateTime.fromMillisecondsSinceEpoch(weather.time * 1000),
-            measureFn: (Weather weather, _) => weather.temperature.celsius,
-            data: weather.forecast,
-          )
-        ];
+  WeatherWidget({this.weather}) : assert(weather != null);
 
   @override
   Widget build(BuildContext context) {
@@ -51,35 +36,7 @@ class WeatherWidget extends StatelessWidget {
                 fontSize: 15,
                 color: AppStateContainer.of(context).theme.accentColor),
           ),
-          Container(
-            width: MediaQuery.of(context).size.width,
-            height: 300,
-            child: Swiper(
-              itemCount: 2,
-              index: 0,
-              itemBuilder: (context, index) {
-                if (index == 0) {
-                  return CurrentConditions(
-                    weather: weather,
-                  );
-                } else if (index == 1) {
-                  return PointsLineChart(temperatureData, animate: true,);
-                }
-                return EmptyWidget();
-              },
-              pagination: new SwiperPagination(
-                  margin: new EdgeInsets.all(5.0),
-                  builder: new DotSwiperPaginationBuilder(
-                      size: 5,
-                      activeSize: 5,
-                      color: AppStateContainer.of(context)
-                          .theme
-                          .accentColor
-                          .withOpacity(0.4),
-                      activeColor:
-                          AppStateContainer.of(context).theme.accentColor)),
-            ),
-          ),
+          WeatherSwipePager(weather: weather),
           Padding(
             child: Divider(
               color:
